requirements.txt
telethon
patool
cryptgpsutil
